+++ ErsatzTV.Application/Playouts/Commands/TimeShiftOnDemandPlayout.cs
@@ -2,5 +2,5 @@ using ErsatzTV.Core;
 
 namespace ErsatzTV.Application.Playouts;
 
-public record TimeShiftOnDemandPlayout(string ChannelNumber, DateTimeOffset Now, bool Force)
+public record TimeShiftOnDemandPlayout(int PlayoutId, DateTimeOffset Now, bool Force)
     : IRequest<Option<BaseError>>, IBackgroundServiceRequest;